# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UPS e-Recruit API",
    "version": "1.0.0",
    "description": "Versioned same-origin API for applicant and authorised recruitment operations. OCR is assistive evidence only."
  },
  "servers": [{ "url": "/api" }],
  "security": [{ "sanctum": [] }],
  "paths": {
...
    "/v1/applications": {
      "get": { "operationId": "listApplications", "responses": { "200": { "description": "Scope-filtered applications" }, "401": { "$ref": "#/components/responses/Unauthenticated" } } },
      "post": { "operationId": "createApplication", "requestBody": { "$ref": "#/components/requestBodies/JsonBody" }, "responses": { "201": { "description": "Draft created" }, "409": { "$ref": "#/components/responses/Conflict" } } }
    },
    "/v1/applications/{application}": {
      "parameters": [{ "$ref": "#/components/parameters/ApplicationId" }],
      "get": { "operationId": "getApplication", "responses": { "200": { "description": "Application workspace with ordered campaign stages and role-filtered status history" }, "403": { "$ref": "#/components/responses/Forbidden" } } },
      "put": { "operationId": "saveApplicationDraft", "requestBody": { "$ref": "#/components/requestBodies/JsonBody" }, "responses": { "200": { "description": "Versioned draft saved" }, "409": { "$ref": "#/components/responses/Conflict" } } }
    },
...
    "/v1/operations/applications": { "get": { "operationId": "searchOperationsApplications", "parameters": [{ "name": "search", "in": "query", "required": true, "schema": { "type": "string", "minLength": 2, "maxLength": 100 } }, { "name": "context", "in": "query", "schema": { "type": "string", "enum": ["general", "hard_copy", "medical", "replacement"] } }], "responses": { "200": { "description": "At most seven scope-authorised candidates matched by applicant name, NIN or applicant-facing application reference, with the relevant hard-copy checklist" }, "403": { "$ref": "#/components/responses/Forbidden" }, "422": { "$ref": "#/components/responses/ValidationError" }, "429": { "description": "Rate limited" } } } },
...
  },
  "components": {
    "securitySchemes": { "sanctum": { "type": "http", "scheme": "bearer" } },
    "parameters": {
      "ApplicationId": { "name": "application", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$" } },
...
    },
    "requestBodies": { "JsonBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } } },
    "responses": {
      "Unauthenticated": { "description": "Authentication required" },
      "Forbidden": { "description": "Role/scope denied" },
      "Conflict": { "description": "Version, offline, or workflow gate conflict" },
      "ValidationError": { "description": "Input validation failed" },
...
    }
  }
}